Write main Gameboard factory functionality
placeShip -> This function places a ship in the grid, when start and
end coordinates are given to the function. Then this function internally
checks if the coordinates given are valid and then proceeds to create a
ship instance with its coordinates and finally place the ship on the grid.

receiveAttack -> This function loops over all the shipsPresent on the grid,
and checks if the coordinates given to the function are equal to any ship cell.
if they are equal, the hit gets displayed in the grid and the hits counter of the
ship that was hit gets incremented. if the coordinates that were passed are not equal
to any ship cell that's present on the grid, then the hit gets represented on the grid
as a missed hit.

areAllShipsSunk -> This function simply parses the array of ships that are present on the
grid. if all the ships are sunk it returns true if only one is still not sunk it returns false.